requirements.txt
pytesseract
pika  # Für die Verbindung zu RabbitMQ
Pillow  # Für die Bildverarbeitung
pdf2image
minio